requirements.txt
pandas
openai
boto3
pytest
streamlit
apache-airflow
great_expectations==0.18.7